+++ ContentPatcher/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Added new tokens:
+  * `HasValue`: check whether any token has a value, like `HasValue:{{spouse}}`.
   * `IsMainPlayer`: whether the player is the main player.
 * Added support for tokens in field keys.
 * Added support for new asset types in the upcoming Stardew Valley 1.4.